requirements.txt
asgiref==3.7.2
Django==5.0.1
djangorestframework==3.14.0
paho-mqtt==1.6.1
pika==1.3.2
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
